requirements.txt
astroid==2.2.5
glfw==1.7.1
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
multipledispatch==0.6.0
numpy==1.16.3
Pillow==6.0.0
pygame==1.9.6
pylint==2.3.1
PyOpenGL==3.1.0
pyrr==0.10.3
six==1.12.0
typed-ast==1.3.5
wrapt==1.11.1
